% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{precompute.for.predictions}
\alias{precompute.for.predictions}
\title{Precompute matrices to use for predictions}
\usage{
precompute.for.predictions(model)
}
\arguments{
\item{model}{A collective matrix factorization model object, for which the
pre-computed matrices will be calculated.}
}
\value{
The same model object, with the pre-calculated matrices inside it.
}
\description{
Pre-computes internal matrices which might be used to speed up
computations on new data in the \link{CMF} and \link{CMF_implicit} models.
This function does not need to be called when passing
`precompute_for_predictions=TRUE`.
}
