% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cmfrec}
\alias{summary.cmfrec}
\title{Get information about factorization model}
\usage{
\method{summary}{cmfrec}(object, ...)
}
\arguments{
\item{object}{An object of class `cmfrec` as returned by functions \link{CMF},
\link{CMF_implicit}, \link{MostPopular}, \link{ContentBased},
\link{OMF_explicit}, \link{OMF_implicit}.}

\item{...}{Extra arguments (not used).}
}
\value{
No return value (information is printed).
}
\description{
Print basic properties of a `cmfrec` object (a base class
encompassing all the models in this package). Same as the `print.cmfrec` function.
}
\seealso{
\link{print.cmfrec}
}
